+++ projects/rxjs-supersets/src/operators/group-delta/group-delta.spec.ts
@@ -46,7 +46,7 @@ describe('groupDelta operator', () => {
   it('should add an element to a group', () => {
     deltaSet.add({id: '1', group: 1});
     expect(results.length).toBe(1);
-    expect(results[0].created.get(1)).toBeDefined();
+    expect(results[0].added.get(1)).toBeDefined();
   });
 
   it('should add multiple elements to a group', () => {
@@ -54,8 +54,8 @@ describe('groupDelta operator', () => {
     results = [];
     deltaSet.add({id: '1b', group: 1});
 
-    // group created in the first update, element added in the second update
-    const group1 = results[0].updated.get(1);
+    // group added in the first update, element added in the second update
+    const group1 = results[0].modified.get(1);
     expect(group1).toBeDefined();
     expect(group1?.map.size).toBe(2);
   });
@@ -65,8 +65,8 @@ describe('groupDelta operator', () => {
     results = [];
     deltaSet.add({id: '1', group: 1});
 
-    // group created in the first update, element updated in the second update
-    const group1 = results[0].updated.get(1);
+    // group added in the first update, element modified in the second update
+    const group1 = results[0].modified.get(1);
     expect(group1).toBeDefined();
     expect(group1?.map.size).toBe(1);
   });
@@ -76,8 +76,8 @@ describe('groupDelta operator', () => {
     results = [];
     deltaSet.add({id: '2', group: 2});
 
-    // group created in the first update, other group created in the second update
-    const group1 = results[0].created.get(2);
+    // group added in the first update, other group added in the second update
+    const group1 = results[0].added.get(2);
     expect(group1).toBeDefined();
     expect(group1?.map.size).toBe(1);
   });
@@ -88,13 +88,13 @@ describe('groupDelta operator', () => {
     results = [];
     deltaSet.add({id: '1b', group: 2});
 
-    // group created in the first update, 
+    // group added in the first update, 
     // element added to th group in the second update
     // element moved to other group in the third update
-    const group1 = results[0].updated.get(1);
+    const group1 = results[0].modified.get(1);
     expect(group1).toBeDefined();
     expect(group1?.map.size).toBe(1);
-    const group2 = results[0].created.get(2);
+    const group2 = results[0].added.get(2);
     expect(group2).toBeDefined();
     expect(group2?.map.size).toBe(1);
   });
@@ -104,12 +104,12 @@ describe('groupDelta operator', () => {
     results = [];
     deltaSet.add({id: '1a', group: 2});
 
-    // group created in the first update, 
+    // group added in the first update, 
     // element moved to other group in the second update
     const group1 = results[0].deleted.get(1);
     expect(group1).toBeDefined();
     expect(group1?.map.size).toBe(0);
-    const group2 = results[0].created.get(2);
+    const group2 = results[0].added.get(2);
     expect(group2).toBeDefined();
     expect(group2?.map.size).toBe(1);
   });