+++ projects/rxjs-supersets/src/operators/start-delta/start-delta.spec.ts
@@ -8,7 +8,7 @@ const element2 = { id: 'element2' };
 const element3 = { id: 'element3' };
 
 describe('startDelta', () => {
-  it('should change the mapDelta.created property to the whole set for the first mapDelta', () => {
+  it('should change the mapDelta.added property to the whole set for the first mapDelta', () => {
     let results: MapDelta<string, IdObject>[] = [];
     const test = new DeltaMap<string, IdObject>();
 
@@ -18,13 +18,13 @@ describe('startDelta', () => {
     const subscription = test.delta$.pipe(startDelta()).subscribe(result => results.push(result));
 
     expect(results.length).toBe(1);
-    expect(results[0].created.size).toBe(2);
+    expect(results[0].added.size).toBe(2);
 
     results = [];
     test.set(element3.id, element3);
 
     expect(results.length).toBe(1);
-    expect(results[0].created.size).toBe(1);
+    expect(results[0].added.size).toBe(1);
 
     subscription.unsubscribe();
   });