+++ projects/rxjs-supersets/src/operators/tap-delta/tap-delta.spec.ts
@@ -13,9 +13,9 @@ describe('tapDelta', () => {
     const test = new DeltaMap<string, IdObject>();
 
     const subscription = test.delta$.pipe(tapDelta({
-      create: element => results.push(`add:${element.id}`),
+      add: element => results.push(`add:${element.id}`),
       delete: element => results.push(`delete:${element.id}`),
-      update: element => results.push(`modify:${element.id}`),
+      modify: element => results.push(`modify:${element.id}`),
       before: () => results.push('before'),
       after: () => results.push('after')
     })).subscribe();